+++ memegen/routes/_utils.py
@@ -1,8 +1,8 @@
-import pprint
 import logging
 from urllib.parse import unquote
 
 import requests
+import background
 from flask import (Response, url_for, render_template, send_file,
                    current_app, request)
 
@@ -10,24 +10,42 @@ from flask import (Response, url_for, render_template, send_file,
 log = logging.getLogger(__name__)
 
 
+def samples(blank=False):
+    """Generate dictionaries of sample image data for template rendering."""
+    for template in sorted(current_app.template_service.all()):
+        path = "_" if blank else template.sample_path
+        url = route('image.get', key=template.key, path=path,
+                    preview=True, watermark='none')
+        yield {
+            'key': template.key,
+            'name': template.name,
+            'url': url,
+        }
+
+
 def route(*args, **kwargs):
     """Unquoted version of Flask's `url_for`."""
+    for key, value in sorted(kwargs.items()):
+        if value is True:
+            kwargs[key] = 'true'
+
     return _secure(unquote(url_for(*args, **kwargs)))
 
 
 def display(title, path, share=False, raw=False, mimetype='image/jpeg'):
     """Render a webpage or raw image based on request."""
-    mimetypes = request.headers.get('Accept', "").split(',')
-    browser = 'text/html' in mimetypes
-
-    img = _format(request, 'share')
-    img_twitter = _format(request, 'share',
-                          width=current_app.config['TWITTER_IMAGE_WIDTH'],
-                          height=current_app.config['TWITTER_IMAGE_HEIGHT'])
-    img_facebook = _format(request, 'share',
-                           width=current_app.config['FACEBOOK_IMAGE_WIDTH'],
-                           height=current_app.config['FACEBOOK_IMAGE_HEIGHT'])
-    url = _format(request, 'width', 'height')
+    img = _format_url(request, 'share')
+    img_twitter = _format_url(
+        request, 'share',
+        width=current_app.config['TWITTER_IMAGE_WIDTH'],
+        height=current_app.config['TWITTER_IMAGE_HEIGHT'],
+    )
+    img_facebook = _format_url(
+        request, 'share',
+        width=current_app.config['FACEBOOK_IMAGE_WIDTH'],
+        height=current_app.config['FACEBOOK_IMAGE_HEIGHT'],
+    )
+    url = _format_url(request, 'width', 'height')
 
     if share:
         log.info("Sharing image on page: %s", img)
@@ -41,18 +59,7 @@ def display(title, path, share=False, raw=False, mimetype='image/jpeg'):
             url=_secure(url),
             config=current_app.config,
         )
-        return html if raw else _nocache(Response(html))
-
-    elif browser:
-        log.info("Embedding image on page: %s", img)
-
-        html = render_template(
-            'image.html',
-            title=title,
-            img=_secure(img),
-            config=current_app.config,
-        )
-        return html if raw else _nocache(Response(html))
+        return html if raw else Response(html)
 
     else:
         log.info("Sending image: %s", path)
@@ -61,11 +68,11 @@ def display(title, path, share=False, raw=False, mimetype='image/jpeg'):
 
 def track(title):
     """Log the requested content, server-side."""
-    tid = current_app.config['GOOGLE_ANALYTICS_TID']
-
-    data = dict(
+    google_url = current_app.config['GOOGLE_ANALYTICS_URL']
+    google_tid = current_app.config['GOOGLE_ANALYTICS_TID']
+    google_data = dict(
         v=1,
-        tid=tid,
+        tid=google_tid,
         cid=request.remote_addr,
 
         t='pageview',
@@ -77,10 +84,24 @@ def track(title):
         ua=request.user_agent.string,
     )
 
-    if tid == 'localhost':
-        log.debug("Analytics data:\n%s", pprint.pformat(data))
-    else:
-        requests.post("http://www.google-analytics.com/collect", data=data)
+    remote_url = current_app.config['REMOTE_TRACKING_URL']
+    remote_data = dict(
+        text=str(title),
+        source='memegen.link',
+        context=unquote(request.url),
+    )
+
+    @background.task
+    def run():
+        if google_tid != 'localhost':
+            response = requests.post(google_url, data=google_data)
+            params = _format_query(google_data, as_string=True)
+            log.debug("Tracking POST: %s %s", response.url, params)
+
+        if remote_url:
+            response = requests.get(remote_url, params=remote_data)
+            log.debug("Tracking GET: %s", response.url)
+    run()
 
 
 def _secure(url):
@@ -91,14 +112,14 @@ def _secure(url):
     return url
 
 
-def _format(req, *skip, **add):
+def _format_url(req, *skip, **add):
     """Get a formatted URL with sanitized query parameters."""
     base = req.base_url
 
     options = {k: v[0] for k, v in dict(req.args).items() if k not in skip}
     options.update(add)
 
-    params = sorted("{}={}".format(k, v) for k, v in options.items())
+    params = _format_query(options)
 
     if params:
         return base + "?{}".format("&".join(params))
@@ -106,9 +127,9 @@ def _format(req, *skip, **add):
         return base
 
 
-def _nocache(response):
-    """Ensure a response is not cached."""
-    response.headers['Cache-Control'] = 'no-cache, no-store, must-revalidate'
-    response.headers['Pragma'] = 'no-cache'
-    response.headers['Expires'] = '0'
-    return response
+def _format_query(options, *, as_string=False):
+    pattern = "{}={!r}" if as_string else "{}={}"
+    pairs = sorted(pattern.format(k, v) for k, v in options.items())
+    if as_string:
+        return ' '.join(pairs)
+    return pairs