+++ memegen/routes/index.py
@@ -1,33 +1,33 @@
 import random
+from pathlib import Path
 
-from flask import Blueprint, render_template, current_app
+from flask import Blueprint, Markup, render_template, current_app
+from flask_cachecontrol import cache_for
+from markdown import markdown
 
-from ._utils import route
+from ._utils import samples
 
 
 blueprint = Blueprint('index-page', __name__)
 
 
 @blueprint.route("/")
-def get_index():
-    template_images = list(_samples(blank=True))
-    sample_images = list(_samples())
+@cache_for(hours=12)
+def get():
+    template_images = list(samples(blank=True))
     return render_template(
         "index.html",
         template_images=template_images,
         default_template=random.choice(template_images)['key'],
-        sample_images=sample_images,
+        readme=_load_readme(),
         config=current_app.config,
     )
 
 
-def _samples(blank=False):
-    """Generate dictionaries of sample image data for template rendering."""
-    for template in sorted(current_app.template_service.all()):
-        path = "_" if blank else template.sample_path
-        url = route('image.get', key=template.key, path=path)
-        yield {
-            'key': template.key,
-            'name': template.name,
-            'url': url,
-        }
+def _load_readme():
+    path = Path(current_app.config['ROOT'], 'README.md')
+    with path.open() as f:
+        text = f.read()
+        content = text.split('<!--content-->')[-1]
+        html = markdown(content, extensions=['tables', 'pymdownx.magiclink'])
+        return Markup(html)