+++ tests/test_api_root.py
@@ -8,15 +8,14 @@ from .utils import load
 def describe_root():
 
     def it_returns_links_and_metadata(client):
-        status, data = load(client.get("/api"))
+        status, data = load(client.get("/api/"))
 
         expect(status) == 200
         expect(data) == {
             'templates': "http://localhost/api/templates/",
             'fonts': "http://localhost/api/fonts/",
             'aliases': "http://localhost/api/aliases/",
-            'magic': "http://localhost/api/magic/",
             'search': "http://localhost/api/search/",
-            'version': "4.5",
+            'version': "5.2",
             'changes': "https://raw.githubusercontent.com/jacebrowning/memegen/master/CHANGELOG.md"
         }